Move subgraphs file
