+++ dev.js
@@ -1,6 +1,9 @@
-import { start as gateway } from "./gateway/server.js";
-import { LOCAL_SUBGRAPH_CONFIG, startSubgraphs } from "./subgraphs.js";
+import { start as gateway } from './gateway/server.js';
+import { LOCAL_SUBGRAPH_CONFIG, startSubgraphs } from './subgraphs/subgraphs.js';
 
+// For local development, we will run a local gateway that will handle
+// composition and configure the ports of the gateway and subgraphs manually.
+// We could use `rover dev` here but that will not work for Stackblitz
 (async () => {
   // start subgraphs in monolith mode
   await startSubgraphs(4001);
